fix an issue with hill drawing where too much downhill would break it
this looks pretty much visually correct and intuitive. the only issue is that hill steepness is comparable within the same course but not between courses, which makes courses with only one or two short hills look much steeper than they actually are.

the only problem is that the hill drawing has become a bit incomprehensible and bad. does it work? yes. can i explain why it works? not really. does it make sense mathematically? certainly don't think so.

all-in-all, i'm not very happy with it overall.