retry solver instances instead of generating new ones when flipping
the comment about *2 removed in this commit was actually wrong. sometimes you need more than 2× as many instances as nsamples because if you flip and regenerate you get a new solver with new skill proc locations and our guess about this one (which was slower, and we thought would be faster, hence the switch) can be wrong _again_, which can lead to using more than 2× as many solver instances. (i'm not explaining this well, but hopefully you get the point.) easiest to re-generate the same solver since now we know for sure which of a/b is actually faster and can re-run accurately.