allow selecting debuff skills multiple times
the implementation of this is VERY NASTY. the assumption that each skill is added only once is very deeply embedded in the frontend (though the simulator is completely fine with it) and skill lists are represented as sets. this creates nasty fake skill ids to get around that. not happy with this at all.